# docs/swagger.yaml
basePath: /
definitions:
...
  models.AutocompleteData:
    properties:
      query:
        type: string
      suggestions:
        items:
          $ref: '#/definitions/models.GuestSuggestion'
        type: array
      total_matches:
        type: integer
    type: object
  models.AutocompleteResponse:
    properties:
      data:
        $ref: '#/definitions/models.AutocompleteData'
      error:
        type: string
      success:
        type: boolean
    type: object
...
  models.ContactType:
    enum:
    - email
    - phone
    - telegram
    - discord
    - twitter
    - linkedin
    - instagram
    - website
    - other
    type: string
    x-enum-varnames:
    - ContactTypeEmail
    - ContactTypePhone
    - ContactTypeTelegram
    - ContactTypeDiscord
    - ContactTypeTwitter
    - ContactTypeLinkedIn
    - ContactTypeInstagram
    - ContactTypeWebsite
    - ContactTypeOther
  models.CreateGuestRequest:
    properties:
      avatar:
        type: string
      contacts:
        items:
          $ref: '#/definitions/models.GuestContact'
        type: array
      metadata:
        additionalProperties: true
        type: object
      name:
        maxLength: 255
        minLength: 1
        type: string
      notes:
        type: string
      short_name:
        type: string
      surname:
        maxLength: 255
        minLength: 1
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - name
    - surname
    type: object
  models.CreateGuestResponse:
    properties:
      data:
        $ref: '#/definitions/models.Guest'
      error:
        type: string
      success:
        type: boolean
    type: object
...
  models.DateRange:
    properties:
      end:
        type: string
      start:
        type: string
    type: object
...
  models.DeleteGuestRequest:
    properties:
      guest_id:
        type: string
    required:
    - guest_id
    type: object
  models.DeleteGuestResponse:
    properties:
      data:
        $ref: '#/definitions/models.GuestDeleteData'
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.GetGuestInfoResponse:
    properties:
      data:
        $ref: '#/definitions/models.GuestInfoData'
      error:
        type: string
      success:
        type: boolean
    type: object
...
  models.Guest:
    properties:
      avatar:
        type: string
      contacts:
        items:
          $ref: '#/definitions/models.GuestContact'
        type: array
      created_at:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      notes:
        type: string
      short_name:
        type: string
      surname:
        type: string
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.GuestContact:
    properties:
      is_primary:
        type: boolean
      label:
        type: string
      type:
        $ref: '#/definitions/models.ContactType'
      value:
        type: string
    type: object
  models.GuestDeleteData:
    properties:
      deleted_at:
        type: string
      guest_id:
        type: string
    type: object
  models.GuestFilters:
    properties:
      created_date_range:
        $ref: '#/definitions/models.DateRange'
      has_contact_type:
        items:
          $ref: '#/definitions/models.ContactType'
        type: array
      search:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  models.GuestInfoData:
    properties:
      guest:
        $ref: '#/definitions/models.Guest'
      stats:
        $ref: '#/definitions/models.GuestStats'
    type: object
  models.GuestListItem:
    properties:
      avatar:
        type: string
      contact_count:
        type: integer
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      notes_preview:
        type: string
      primary_email:
        type: string
      short_name:
        type: string
      surname:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  models.GuestSortOptions:
    properties:
      field:
        type: string
      order:
        type: string
    type: object
  models.GuestStats:
    properties:
      last_appearance:
        type: string
      total_shows:
        type: integer
      upcoming_shows:
        type: integer
    type: object
  models.GuestSuggestion:
    properties:
      avatar:
        type: string
      display_name:
        type: string
      id:
        type: string
      match_score:
        type: number
      name:
        type: string
      primary_contact:
        $ref: '#/definitions/models.GuestContact'
      short_name:
        type: string
      surname:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  models.ListGuestsData:
    properties:
      guests:
        items:
          $ref: '#/definitions/models.GuestListItem'
        type: array
      pagination:
        $ref: '#/definitions/models.PaginationResponse'
    type: object
  models.ListGuestsRequest:
    properties:
      filters:
        $ref: '#/definitions/models.GuestFilters'
      pagination:
        $ref: '#/definitions/models.PaginationOptions'
      sort:
        $ref: '#/definitions/models.GuestSortOptions'
    type: object
  models.ListGuestsResponse:
    properties:
      data:
        $ref: '#/definitions/models.ListGuestsData'
      error:
        type: string
      success:
        type: boolean
    type: object
...
  models.PaginationOptions:
    properties:
      limit:
        type: integer
      page:
        type: integer
      sort:
        type: string
    type: object
  models.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.UpdateGuestRequest:
    properties:
      avatar:
        type: string
      contacts:
        items:
          $ref: '#/definitions/models.GuestContact'
        type: array
      guest_id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      notes:
        type: string
      short_name:
        type: string
      surname:
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - guest_id
    type: object
  models.UpdateGuestResponse:
    properties:
      data:
        $ref: '#/definitions/models.Guest'
      error:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A Go-based microservice for planning and scheduling YouTube live streams.
    Helps content creators organize their streaming schedule and manage content.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: St. Planer - YouTube Stream Planner API
  version: "1.0"
paths:
...
  /api/v1/guest/autocomplete:
    get:
      description: Get guest suggestions for autocomplete functionality
      parameters:
      - description: Search query (minimum 2 characters)
        in: query
        name: q
        required: true
        type: string
      - description: 'Maximum results (default: 10, max: 50)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AutocompleteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Guest autocomplete search
      tags:
      - guests
  /api/v1/guest/delete:
    delete:
      consumes:
      - application/json
      description: Delete a guest from the system
      parameters:
      - description: Guest deletion data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteGuestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteGuestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete guest
      tags:
      - guests
  /api/v1/guest/info/{guest_id}:
    get:
      description: Get detailed information about a specific guest
      parameters:
      - description: Guest ID
        in: path
        name: guest_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetGuestInfoResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get guest details
      tags:
      - guests
  /api/v1/guest/list:
    post:
      consumes:
      - application/json
      description: Get paginated list of guests with filtering and sorting options
      parameters:
      - description: Guest list filters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ListGuestsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListGuestsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List guests with filtering
      tags:
      - guests
  /api/v1/guest/new:
    post:
      consumes:
      - application/json
      description: Create a new guest with contact information and notes
      parameters:
      - description: Guest creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateGuestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateGuestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create new guest
      tags:
      - guests
  /api/v1/guest/update:
    put:
      consumes:
      - application/json
      description: Update existing guest with new information
      parameters:
      - description: Guest update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateGuestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateGuestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update guest information
      tags:
      - guests
...
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"